requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
boto3==1.39.3
botocore==1.39.3
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
fastapi==0.116.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.5.5.64
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
s3transfer==0.13.0
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
timm==0.5.4
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.47.0
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
